pyproject.toml
[project]
name = "PBIS"
version = "0.1.0"
description = ""
authors = [
    {name = "qasim0900",email = "qasim_kamran2007@yahoo.com"}
]
package-mode = false
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "whitenoise (>=6.11.0,<7.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]

[project.optional-dependencies]
test = [
    "pytest (>=8.0.0,<9.0.0)",
    "pytest-django (>=4.5.2,<5.0.0)",
    "pytest-cov (>=4.0.0,<5.0.0)",
    "factory-boy (>=3.3.0,<4.0.0)",
    "model-bakery (>=1.12.0,<2.0.0)",
    "django-test-plus (>=2.2.2,<3.0.0)",
    "requests-mock (>=1.11.0,<2.0.0)",
    "freezegun (>=1.4.0,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"